# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BacklogMCP
  description: "Backlog SaaS\u3092Model Context Protocol (MCP)\u7D4C\u7531\u3067\u64CD\
    \u4F5C\u3059\u308B\u305F\u3081\u306EAPI"
  version: 0.1.0
paths:
...
  /api/bulk/status:
    post:
      tags:
      - bulk_operations
      summary: Bulk Update Status
      description: "\u8907\u6570\u30C1\u30B1\u30C3\u30C8\u306E\u30B9\u30C6\u30FC\u30BF\
        \u30B9\u3092\u4E00\u62EC\u66F4\u65B0\u3059\u308B\u30A8\u30F3\u30C9\u30DD\u30A4\
        \u30F3\u30C8\n\nArgs:\n    request: \u4E00\u62EC\u66F4\u65B0\u30EA\u30AF\u30A8\
        \u30B9\u30C8\n    bulk_service: \u4E00\u62EC\u64CD\u4F5C\u30B5\u30FC\u30D3\
        \u30B9\uFF08\u4F9D\u5B58\u6027\u6CE8\u5165\uFF09\n\nReturns:\n    \u51E6\u7406\
        \u7D50\u679C\u306E\u7D71\u8A08\u60C5\u5831"
      operationId: bulk_update_status
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkStatusUpdateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Bulk Update Status
        '404':
          description: Not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/bulk/assignee:
    post:
      tags:
      - bulk_operations
      summary: Bulk Update Assignee
      description: "\u8907\u6570\u30C1\u30B1\u30C3\u30C8\u306E\u62C5\u5F53\u8005\u3092\
        \u4E00\u62EC\u66F4\u65B0\u3059\u308B\u30A8\u30F3\u30C9\u30DD\u30A4\u30F3\u30C8\
        \n\nArgs:\n    request: \u4E00\u62EC\u66F4\u65B0\u30EA\u30AF\u30A8\u30B9\u30C8\
        \n    bulk_service: \u4E00\u62EC\u64CD\u4F5C\u30B5\u30FC\u30D3\u30B9\uFF08\
        \u4F9D\u5B58\u6027\u6CE8\u5165\uFF09\n\nReturns:\n    \u51E6\u7406\u7D50\u679C\
        \u306E\u7D71\u8A08\u60C5\u5831"
      operationId: bulk_update_assignee
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkAssigneeUpdateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Bulk Update Assignee
        '404':
          description: Not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/bulk/priority:
    post:
      tags:
      - bulk_operations
      summary: Bulk Update Priority
      description: "\u8907\u6570\u30C1\u30B1\u30C3\u30C8\u306E\u512A\u5148\u5EA6\u3092\
        \u4E00\u62EC\u66F4\u65B0\u3059\u308B\u30A8\u30F3\u30C9\u30DD\u30A4\u30F3\u30C8\
        \n\nArgs:\n    request: \u4E00\u62EC\u66F4\u65B0\u30EA\u30AF\u30A8\u30B9\u30C8\
        \n    bulk_service: \u4E00\u62EC\u64CD\u4F5C\u30B5\u30FC\u30D3\u30B9\uFF08\
        \u4F9D\u5B58\u6027\u6CE8\u5165\uFF09\n\nReturns:\n    \u51E6\u7406\u7D50\u679C\
        \u306E\u7D71\u8A08\u60C5\u5831"
      operationId: bulk_update_priority
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkPriorityUpdateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Bulk Update Priority
        '404':
          description: Not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/bulk/milestone:
    post:
      tags:
      - bulk_operations
      summary: Bulk Update Milestone
      description: "\u8907\u6570\u30C1\u30B1\u30C3\u30C8\u306E\u30DE\u30A4\u30EB\u30B9\
        \u30C8\u30FC\u30F3\u3092\u4E00\u62EC\u66F4\u65B0\u3059\u308B\u30A8\u30F3\u30C9\
        \u30DD\u30A4\u30F3\u30C8\n\nArgs:\n    request: \u4E00\u62EC\u66F4\u65B0\u30EA\
        \u30AF\u30A8\u30B9\u30C8\n    bulk_service: \u4E00\u62EC\u64CD\u4F5C\u30B5\
        \u30FC\u30D3\u30B9\uFF08\u4F9D\u5B58\u6027\u6CE8\u5165\uFF09\n\nReturns:\n\
        \    \u51E6\u7406\u7D50\u679C\u306E\u7D71\u8A08\u60C5\u5831"
      operationId: bulk_update_milestone
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkMilestoneUpdateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Bulk Update Milestone
        '404':
          description: Not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/bulk/category:
    post:
      tags:
      - bulk_operations
      summary: Bulk Update Category
      description: "\u8907\u6570\u30C1\u30B1\u30C3\u30C8\u306E\u30AB\u30C6\u30B4\u30EA\
        \u3092\u4E00\u62EC\u66F4\u65B0\u3059\u308B\u30A8\u30F3\u30C9\u30DD\u30A4\u30F3\
        \u30C8\n\nArgs:\n    request: \u4E00\u62EC\u66F4\u65B0\u30EA\u30AF\u30A8\u30B9\
        \u30C8\n    bulk_service: \u4E00\u62EC\u64CD\u4F5C\u30B5\u30FC\u30D3\u30B9\
        \uFF08\u4F9D\u5B58\u6027\u6CE8\u5165\uFF09\n\nReturns:\n    \u51E6\u7406\u7D50\
        \u679C\u306E\u7D71\u8A08\u60C5\u5831"
      operationId: bulk_update_category
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkCategoryUpdateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Bulk Update Category
        '404':
          description: Not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/bulk/delete:
    post:
      tags:
      - bulk_operations
      summary: Bulk Delete Issues
      description: "\u8907\u6570\u30C1\u30B1\u30C3\u30C8\u3092\u4E00\u62EC\u524A\u9664\
        \u3059\u308B\u30A8\u30F3\u30C9\u30DD\u30A4\u30F3\u30C8\n\nArgs:\n    request:\
        \ \u4E00\u62EC\u66F4\u65B0\u30EA\u30AF\u30A8\u30B9\u30C8\n    bulk_service:\
        \ \u4E00\u62EC\u64CD\u4F5C\u30B5\u30FC\u30D3\u30B9\uFF08\u4F9D\u5B58\u6027\
        \u6CE8\u5165\uFF09\n\nReturns:\n    \u51E6\u7406\u7D50\u679C\u306E\u7D71\u8A08\
        \u60C5\u5831"
      operationId: bulk_delete_issues
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkUpdateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Bulk Delete Issues
        '404':
          description: Not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    BulkAssigneeUpdateRequest:
      properties:
        issue_ids:
          items:
            type: string
          type: array
          title: Issue Ids
        assignee_id:
          type: integer
          title: Assignee Id
      type: object
      required:
      - issue_ids
      - assignee_id
      title: BulkAssigneeUpdateRequest
      description: "\u62C5\u5F53\u8005\u4E00\u62EC\u66F4\u65B0\u30EA\u30AF\u30A8\u30B9\
        \u30C8"
    BulkCategoryUpdateRequest:
      properties:
        issue_ids:
          items:
            type: string
          type: array
          title: Issue Ids
        category_id:
          type: integer
          title: Category Id
      type: object
      required:
      - issue_ids
      - category_id
      title: BulkCategoryUpdateRequest
      description: "\u30AB\u30C6\u30B4\u30EA\u4E00\u62EC\u66F4\u65B0\u30EA\u30AF\u30A8\
        \u30B9\u30C8"
    BulkMilestoneUpdateRequest:
      properties:
        issue_ids:
          items:
            type: string
          type: array
          title: Issue Ids
        milestone_id:
          type: integer
          title: Milestone Id
      type: object
      required:
      - issue_ids
      - milestone_id
      title: BulkMilestoneUpdateRequest
      description: "\u30DE\u30A4\u30EB\u30B9\u30C8\u30FC\u30F3\u4E00\u62EC\u66F4\u65B0\
        \u30EA\u30AF\u30A8\u30B9\u30C8"
    BulkPriorityUpdateRequest:
      properties:
        issue_ids:
          items:
            type: string
          type: array
          title: Issue Ids
        priority_id:
          type: integer
          title: Priority Id
      type: object
      required:
      - issue_ids
      - priority_id
      title: BulkPriorityUpdateRequest
      description: "\u512A\u5148\u5EA6\u4E00\u62EC\u66F4\u65B0\u30EA\u30AF\u30A8\u30B9\
        \u30C8"
    BulkStatusUpdateRequest:
      properties:
        issue_ids:
          items:
            type: string
          type: array
          title: Issue Ids
        status_id:
          type: integer
          title: Status Id
      type: object
      required:
      - issue_ids
      - status_id
      title: BulkStatusUpdateRequest
      description: "\u30B9\u30C6\u30FC\u30BF\u30B9\u4E00\u62EC\u66F4\u65B0\u30EA\u30AF\
        \u30A8\u30B9\u30C8"
    BulkUpdateRequest:
      properties:
        issue_ids:
          items:
            type: string
          type: array
          title: Issue Ids
      type: object
      required:
      - issue_ids
      title: BulkUpdateRequest
      description: "\u4E00\u62EC\u66F4\u65B0\u30EA\u30AF\u30A8\u30B9\u30C8"
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError